requirements.txt
aiohttp==3.0.5
astroid==1.6.1
async-timeout==2.0.0
attrs==17.4.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
discord.py==0.16.12
google-api-python-client==1.6.5
gspread==0.6.2
httplib2==0.10.3
idna==2.6
idna-ssl==1.0.0
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.1.1
mccabe==0.6.1
multidict==4.1.0
oauth2client==4.1.2
pip-chill==0.1.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
pylint==1.8.2
requests==2.18.4
rsa==3.4.2
safygiphy==1.1.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
websockets==3.4
wrapt==1.10.11
yarl==1.1.1
asyncpg